// Frontend/src/api/models/IdeaUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Idea update schema
 */
export type IdeaUpdate = {
    demand_type?: (string | null);
    fast_track?: (boolean | null);
    fast_track_reason?: (string | null);
    start_date?: (string | null);
    end_date?: (string | null);
    description?: (string | null);
    responsible_unit?: (string | null);
    assessment_notes?: (string | null);
    owner_department?: (string | null);
    status?: (string | null);
};